+++ internal/service/cloudapi/kvmvm/models.go
@@ -74,7 +74,7 @@ type DiskRecord struct {
 	SepID               int                    `json:"sepId"` // NOTE: absent from compute/get output
 	Shareable           bool                   `json:"shareable"`
 	SizeMax             int                    `json:"sizeMax"`
-	SizeUsed            int                    `json:"sizeUsed"` // sum over all snapshots of this disk to report total consumed space
+	SizeUsed            float64                `json:"sizeUsed"` // sum over all snapshots of this disk to report total consumed space
 	Snapshots           []SnapshotRecord       `json:"snapshots"`
 	Status              string                 `json:"status"`
 	TechStatus          string                 `json:"techStatus"`
